Refactor Tokenizer to use the delegate pattern
