+++ lib/simple-html-tokenizer/tokenizer.js
@@ -5,18 +5,13 @@ import {
   Chars,
   Comment
 } from './tokens';
+import EventedTokenizer from './evented-tokenizer';
 
 function Tokenizer(entityParser) {
   this.entityParser = entityParser;
+  this.tokenizer = new EventedTokenizer(this);
 
-  this.input = null;
-  this.state = null;
   this.token = null;
-
-  this.char = -1;
-  this.line = -1;
-  this.column = -1;
-
   this.startLine = -1;
   this.startColumn = -1;
 
@@ -24,76 +19,27 @@ function Tokenizer(entityParser) {
 }
 
 Tokenizer.prototype = {
+  tokenize: function(input) {
+    return this.tokenizer.tokenize(input);
+  },
+
+  tokenizePart: function(input) {
+    return this.tokenizer.tokenizePart(input);
+  },
+
+  tokenizeEOF: function() {
+    return this.tokenizer.tokenizeEOF();
+  },
+
   reset: function() {
-    this.input = '';
-    this.state = 'data';
     this.token = null;
-
-    this.char = 0;
-    this.line = 1;
-    this.column = 0;
-
     this.startLine = 1;
     this.startColumn = 0;
   },
 
-  tokenize: function(input) {
-    this.reset();
-
-    var tokens = this.tokenizePart(input);
-    var trailingToken = this.tokenizeEOF();
-
-    if (trailingToken) {
-      tokens.push(trailingToken);
-    }
-
-    return tokens
-  },
-
-  tokenizePart: function(input) {
-    this.input += preprocessInput(input);
-
-    var tokens = [];
-
-    while (true) {
-      var token = this.lex();
-
-      if (token) {
-        tokens.push(token);
-      } else {
-        break;
-      }
-    }
-
-    return tokens;
-  },
-
-  tokenizeEOF: function() {
-    return this.emitToken();
-  },
-
-  lex: function() {
-    while (this.char < this.input.length) {
-      var char = this.input.charAt(this.char++);
-      if (char) {
-        if (char === "\n") {
-          this.line++;
-          this.column = 0;
-        } else {
-          this.column++;
-        }
-
-        var token = this.states[this.state].call(this, char);
-        if (token) {
-          return token;
-        }
-      }
-    }
-  },
-
   addLocInfo: function(_endLine, _endColumn) {
-    var endLine = (_endLine === undefined) ? this.line : _endLine;
-    var endColumn = (_endColumn === undefined) ? this.column : _endColumn;
+    var endLine = (_endLine === undefined) ? this.tokenizer.line : _endLine;
+    var endColumn = (_endColumn === undefined) ? this.tokenizer.column : _endColumn;
 
     this.token.loc = {
       start: {
@@ -113,7 +59,6 @@ Tokenizer.prototype = {
   createTag: function(Type, char) {
     var lastToken = this.token;
     this.token = new Type(char);
-    this.state = 'tagName';
     return lastToken;
   },
 
@@ -128,7 +73,7 @@ Tokenizer.prototype = {
   createAttribute: function(char) {
     this._currentAttribute = [char.toLowerCase(), "", null];
     this.token.attributes.push(this._currentAttribute);
-    this.state = 'attributeName';
+    this.tokenizer.state = 'attributeName';
   },
 
   addToAttributeName: function(char) {
@@ -156,7 +101,7 @@ Tokenizer.prototype = {
   createComment: function() {
     var lastToken = this.token;
     this.token = new Comment();
-    this.state = 'commentStart';
+    this.tokenizer.state = 'commentStart';
     return lastToken;
   },
 
@@ -177,11 +122,11 @@ Tokenizer.prototype = {
   emitData: function() {
     var token = this.token;
     if (token) {
-      this.addLocInfo(this.line, this.column - 1);
+      this.addLocInfo(this.tokenizer.line, this.tokenizer.column - 1);
     }
 
     this.token = null;
-    this.state = 'tagOpen';
+    this.tokenizer.state = 'tagOpen';
 
     return token;
   },
@@ -194,7 +139,7 @@ Tokenizer.prototype = {
     }
 
     this.token = null;
-    this.state = 'data';
+    this.tokenizer.state = 'data';
 
     return token;
   },
@@ -208,220 +153,7 @@ Tokenizer.prototype = {
   },
 
   consumeCharRef: function() {
-    return this.entityParser.parse(this);
-  },
-
-  states: {
-    data: function(char) {
-      if (char === "<") {
-        var chars = this.emitData();
-        return chars;
-      } else if (char === "&") {
-        this.addData(this.consumeCharRef() || "&");
-      } else {
-        this.addData(char);
-      }
-    },
-
-    tagOpen: function(char) {
-      if (char === "!") {
-        this.state = 'markupDeclaration';
-      } else if (char === "/") {
-        this.state = 'endTagOpen';
-      } else if (isAlpha(char)) {
-        return this.createTag(StartTag, char.toLowerCase());
-      }
-    },
-
-    markupDeclaration: function(char) {
-      if (char === "-" && this.input.charAt(this.char) === "-") {
-        this.char++;
-        this.createComment();
-      }
-    },
-
-    commentStart: function(char) {
-      if (char === "-") {
-        this.state = 'commentStartDash';
-      } else if (char === ">") {
-        return this.emitToken();
-      } else {
-        this.addToComment(char);
-        this.state = 'comment';
-      }
-    },
-
-    commentStartDash: function(char) {
-      if (char === "-") {
-        this.state = 'commentEnd';
-      } else if (char === ">") {
-        return this.emitToken();
-      } else {
-        this.addToComment("-");
-        this.state = 'comment';
-      }
-    },
-
-    comment: function(char) {
-      if (char === "-") {
-        this.state = 'commentEndDash';
-      } else {
-        this.addToComment(char);
-      }
-    },
-
-    commentEndDash: function(char) {
-      if (char === "-") {
-        this.state = 'commentEnd';
-      } else {
-        this.addToComment("-" + char);
-        this.state = 'comment';
-      }
-    },
-
-    commentEnd: function(char) {
-      if (char === ">") {
-        return this.emitToken();
-      } else {
-        this.addToComment("--" + char);
-        this.state = 'comment';
-      }
-    },
-
-    tagName: function(char) {
-      if (isSpace(char)) {
-        this.state = 'beforeAttributeName';
-      } else if (char === "/") {
-        this.state = 'selfClosingStartTag';
-      } else if (char === ">") {
-        return this.emitToken();
-      } else {
-        this.addToTagName(char);
-      }
-    },
-
-    beforeAttributeName: function(char) {
-      if (isSpace(char)) {
-        return;
-      } else if (char === "/") {
-        this.state = 'selfClosingStartTag';
-      } else if (char === ">") {
-        return this.emitToken();
-      } else {
-        this.createAttribute(char);
-      }
-    },
-
-    attributeName: function(char) {
-      if (isSpace(char)) {
-        this.state = 'afterAttributeName';
-      } else if (char === "/") {
-        this.state = 'selfClosingStartTag';
-      } else if (char === "=") {
-        this.state = 'beforeAttributeValue';
-      } else if (char === ">") {
-        return this.emitToken();
-      } else {
-        this.addToAttributeName(char);
-      }
-    },
-
-    afterAttributeName: function(char) {
-      if (isSpace(char)) {
-        return;
-      } else if (char === "/") {
-        this.state = 'selfClosingStartTag';
-      } else if (char === "=") {
-        this.state = 'beforeAttributeValue';
-      } else if (char === ">") {
-        return this.emitToken();
-      } else {
-        this.finalizeAttributeValue();
-        this.createAttribute(char);
-      }
-    },
-
-    beforeAttributeValue: function(char) {
-      if (isSpace(char)) {
-        return;
-      } else if (char === '"') {
-        this.state = 'attributeValueDoubleQuoted';
-        this.markAttributeQuoted(true);
-      } else if (char === "'") {
-        this.state = 'attributeValueSingleQuoted';
-        this.markAttributeQuoted(true);
-      } else if (char === ">") {
-        return this.emitToken();
-      } else {
-        this.state = 'attributeValueUnquoted';
-        this.markAttributeQuoted(false);
-        this.addToAttributeValue(char);
-      }
-    },
-
-    attributeValueDoubleQuoted: function(char) {
-      if (char === '"') {
-        this.finalizeAttributeValue();
-        this.state = 'afterAttributeValueQuoted';
-      } else if (char === "&") {
-        this.addToAttributeValue(this.consumeCharRef('"') || "&");
-      } else {
-        this.addToAttributeValue(char);
-      }
-    },
-
-    attributeValueSingleQuoted: function(char) {
-      if (char === "'") {
-        this.finalizeAttributeValue();
-        this.state = 'afterAttributeValueQuoted';
-      } else if (char === "&") {
-        this.addToAttributeValue(this.consumeCharRef("'") || "&");
-      } else {
-        this.addToAttributeValue(char);
-      }
-    },
-
-    attributeValueUnquoted: function(char) {
-      if (isSpace(char)) {
-        this.finalizeAttributeValue();
-        this.state = 'beforeAttributeName';
-      } else if (char === "&") {
-        this.addToAttributeValue(this.consumeCharRef(">") || "&");
-      } else if (char === ">") {
-        return this.emitToken();
-      } else {
-        this.addToAttributeValue(char);
-      }
-    },
-
-    afterAttributeValueQuoted: function(char) {
-      if (isSpace(char)) {
-        this.state = 'beforeAttributeName';
-      } else if (char === "/") {
-        this.state = 'selfClosingStartTag';
-      } else if (char === ">") {
-        return this.emitToken();
-      } else {
-        this.char--;
-        this.state = 'beforeAttributeName';
-      }
-    },
-
-    selfClosingStartTag: function(char) {
-      if (char === ">") {
-        this.selfClosing();
-        return this.emitToken();
-      } else {
-        this.char--;
-        this.state = 'beforeAttributeName';
-      }
-    },
-
-    endTagOpen: function(char) {
-      if (isAlpha(char)) {
-        this.createTag(EndTag, char.toLowerCase());
-      }
-    }
+    return this.entityParser.parse(this.tokenizer);
   }
 };
